Got a nice working algorithm for responsive download strategy

+++ cmd/torrentfs/main.go
@@ -29,7 +29,7 @@ var (
 	disableTrackers = flag.Bool("disableTrackers", false, "disables trackers")
 	testPeer        = flag.String("testPeer", "", "the address for a test peer")
 	httpAddr        = flag.String("httpAddr", "localhost:0", "HTTP server bind address")
-	readaheadBytes  = flag.Int("readaheadBytes", 10*1024*1024, "bytes to readahead in each torrent from the last read piece")
+	readaheadBytes  = flag.Int64("readaheadBytes", 10*1024*1024, "bytes to readahead in each torrent from the last read piece")
 	testPeerAddr    *net.TCPAddr
 	listenAddr      = flag.String("listenAddr", ":6882", "incoming connection address")
 )
